hdl/modules/wb_un_cross/wbgen/*: (1/5) update wbgen files
Add switching clock output enable. This effectively
starts the switch counter from 0 to the max value.
This is necessary as the RFFE controller needs to
be in a known state before the first switching clock
edge.

This is accomplished by disabling the switching clock,
commanding the RFFE controller to the direct switch state
and then enabling the switching clock