hdl/modules/cross_uncross_core/*: disable half switching clock generation
Now, the FPGA internals use the same clock as output to the RFFEs.
Previously, we had to use half the RFFE switching clock inside
the FPGA, as the RFFE employed a JK flip-flop scheme, in which the
switching clock was re-generated locally. So, the positive edge
was like a trigger to change the state of the RFFE switches, giving
a frequency half of the incoming signal.