modules/cic/cic_decim.v: fix multiple valid_o signals when no valid_i
This CIC module employs a dynamic decimation signal
to control the decimation and COMB stages.

In previous implementations, on receiving a
act_out_i, which is equivalent to a decimation strobe,
a valid_o would be generated regardless of the
valid_i.

This patch fixes it, by clearing the internal valid
signal when act_out_i is received. This causes, the
valid_o signal to be generated only when new
samples (valid_i) is received.