modules/*/*strober.vhd: reimplement to fix late detection and corner case
This commit aims to fix two issues with the
previous implementation.

1) Late count detection. With the previous
implementation we needed two clock cycles
to output the decimation strobe. One to decrement
the counter and another one to detect the 0.

2) Corner decrement case. With previous
implementation we assumed that on the same clock
cycle the counter was zeroed, we would have a
valid_i signal. In this case, we need to reset the
counter with one count less than the ratio_i.
However, if the counter was zeroed, but valid_i
was not asserted, we need to reset the counter to its
full ratio_i.