modules/delta_sigma: fix valid/ce signals logic by using "valid" ports
Previously some modules didn't have a "valid" input
port, so we either didn't use a valid signal or ANDed
the CE with Valid signal and assign it to the CE component
port (which is a bad thing to do in general).

Now, having implemented the "valid" semantics in the other
modules, we assign the input valid signals to the "valid"
inputs.