modules/*/*calc_core: fix sw scheme dsp signal order
Up until this point we were treating the DSP
signals as ch0, ch1, ch2 and ch3, so as to reflect
the signal that were input to the ADC. So, ch0 is
the signal coming from ADC ch0, and so on.
However, in the position calculation application,
we have the switching scheme between ADC channels
ch0 <-> ch1 and ch2 <-> ch3 that are in fact
connected to the analog signals in the following
way:

ch0 <-> A
ch1 <-> C
ch2 <-> B
ch3 <-> D

Next, we use the position calculation formula
considering the analog signal convention (A, B, C,
D) as this is how the information is coded. This
have the implication that we must input the A, B,
C, D in the correct order for the position_calc
module, but this still uses ch0, ch1, ch2, ch3
convention.

So, to fix this, we must change the order of the
input signal to match the analog <-> digital
signal domains using: ch0 as channel A, ch1 as
channel C, and so on according to the scheme
above.