wb_position_calc_core/wbgen: add new FIFO register for MONIT amp/pos
This is a new interface for safely reading
MONIT. AMP/POS data from the core.

Now, the bus needs to just read from the same
register and it will safely update the monit value
to the next one. Caution must still be taken with
slow clients, as the fifo will become full and only
after this it will start to output the most recent value.

It would be better to employ a circular buffer here,
but this should suffice for now, as 10Hz is pretty easy
to achieve in software.