modules/wb_bpm_swap/wbgen: remove div_clk external load register mode
This mode was introduced in commit
20c7091fc so to allow the bpm_swap module
to detect a new change in div_clk value
and reset the clock swap counter accordingly.

This is ok, as long as we don't reset
FS domain expecting the register to retain
its value, as it happens with the other
registers (without LOAD_EXT).

The problem arises as a reset can happen if
the MMCM unlocks, typically because the user
changed the Si571 clock frequency.

In this case, the MMCM will unlock, resetting
FS clock domain along with div_clk register.
The other registers will be reset only on the FS side,
as the Wishbone block still retains the old value
until a sys_rst happens. When the fs_rst ends,
the other register values will take place.

Now, we went back to the traditional behavior
and will check if we need to reset the clock
counter by comparing the previous div_clk with
a new div_clk. If these are different, reset the
counter.