testbench/cic: use integer values in input data of CIC simulation
To allow better control of simulation, the input data file which
is generated by Octave script 'cic_samplegen.m' is now filled with
integer values instead of real values. Among the errors found in
the approach using real values was the truncation of input data
due to lack of decimals of real values in text representation using
scientific notation.