modules/position_calc: use regular data sync method for TbT CIC
Previously, we were using an input signal to reset
the TbT CIC. This is not optimal, as there was
the CIC sync feature already in place, but it required
a squared wave and not a pulse.

Now, we changed it to use the squared wave (or tag
as we call it). So, an external module must convert
the pulse into the appropriate signal.