modules/wb_position_calc: generate TbT tag with counter instead of trigger
Previously, we needed a trigger coming at the exact
the TbT rate to keep the TbT CIC properly synched.
Failing to do so would keep the CIC in a reset state.

Now, we generate the tag with a counter and only use
the trigger to synchronize the counter with it, so
even without the trigger TbT would still work.