Added testbench for comparing implementations
Added testbench for comparing the Verilog implementation with the VHDL.
For this to work, I created a copy of the Verilog core to be able to import
it on the testbench. I also had to change `positive` to `natural` on the
VHDL core due to modelsim assuming that a `positive` cannot be zero.